requirements.txt
groq 
pypdf
dotenv 
redis 
rq 
psycopg2-binary
langgraph 
langchain_tavily
wave
openai-whisper
numpy 
sounddevice
fastapi 
uvicorn 
pydantic  
bcrypt 
langchain_groq